requirements.txt
requests>=2,<3
google-cloud-firestore>=2
